requirements.txt
# Google Docs API Client Library Dependencies

# Core Google API libraries
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# HTTP and authentication
requests>=2.31.0
urllib3>=1.26.0
httplib2>=0.22.0

# OAuth2 and authentication utilities
oauthlib>=3.2.0
requests-oauthlib>=1.3.0

# JSON and data handling
certifi>=2023.7.22

# Caching (optional but recommended for performance)
cachetools>=5.3.0

# Data processing
pandas>=1.5.0

# Type checking support (optional, for development)
# Uncomment if you want type checking support
# mypy>=1.5.0
# types-requests>=2.31.0

# Development and testing dependencies (optional)
# Uncomment for development setup
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.0.0
# isort>=5.12.0

# Documentation generation (optional)
# Uncomment for documentation generation
# sphinx>=7.1.0
# sphinx-rtd-theme>=1.3.0
